requirements.txt
streamlit
pandas
python-dotenv
black
isort
beautifulsoup4